DVE: Fix signed-unsigned comparison warnings

+++ divine/dve/symtab.h
@@ -115,7 +115,7 @@ struct Symbol {
 
     template< typename T >
     inline void setval( char * place, int value, ErrorState &err ) {
-        if ( value != (T)value )
+        if ( value != (int)(T)value )
             err |= ErrorState::i_overflow;
         *reinterpret_cast< T * >( place ) = value;
     }
